Rename internal folder to pkg
We rename the folder to "pkg" instead of "internal"
so that we will be able to use the helped functions
of different modules outside of commatrix.
