+++ pkg/ss/ss.go
@@ -1,7 +1,6 @@
 package ss
 
 import (
-	"encoding/json"
 	"fmt"
 	"regexp"
 	"strings"
@@ -9,11 +8,11 @@ import (
 
 	corev1 "k8s.io/api/core/v1"
 
-	"github.com/liornoy/node-comm-lib/internal/client"
-	"github.com/liornoy/node-comm-lib/internal/consts"
-	"github.com/liornoy/node-comm-lib/internal/debug"
-	"github.com/liornoy/node-comm-lib/internal/nodes"
-	"github.com/liornoy/node-comm-lib/internal/types"
+	"github.com/liornoy/node-comm-lib/pkg/client"
+	"github.com/liornoy/node-comm-lib/pkg/consts"
+	"github.com/liornoy/node-comm-lib/pkg/debug"
+	"github.com/liornoy/node-comm-lib/pkg/nodes"
+	"github.com/liornoy/node-comm-lib/pkg/types"
 )
 
 const (
@@ -60,11 +59,11 @@ func CreateComDetailsFromNode(cs *client.ClientSet, node *corev1.Node) ([]types.
 	ssOutFilteredTCP := filterStrings(tcpSSFilterFn, splitByLines(ssOutTCP))
 	ssOutFilteredUDP := filterStrings(udpSSFilterFn, splitByLines(ssOutUDP))
 
-	tcpComDetails, err := toComDetails(debugPod, ssOutFilteredTCP, "TCP", node)
+	tcpComDetails, err := toComDetails(ssOutFilteredTCP, "TCP", node)
 	if err != nil {
 		return nil, err
 	}
-	udpComDetails, err := toComDetails(debugPod, ssOutFilteredUDP, "UDP", node)
+	udpComDetails, err := toComDetails(ssOutFilteredUDP, "UDP", node)
 	if err != nil {
 		return nil, err
 	}
@@ -81,12 +80,12 @@ func splitByLines(bytes []byte) []string {
 	return strings.Split(str, "\n")
 }
 
-func toComDetails(debugPod *debug.DebugPod, ssOutput []string, protocol string, node *corev1.Node) ([]types.ComDetails, error) {
+func toComDetails(ssOutput []string, protocol string, node *corev1.Node) ([]types.ComDetails, error) {
 	res := make([]types.ComDetails, 0)
 	nodeRoles := nodes.GetRoles(node)
 
 	for _, ssEntry := range ssOutput {
-		cd, err := parseComDetail(debugPod, ssEntry)
+		cd, err := parseComDetail(ssEntry)
 		if err != nil {
 			return nil, err
 		}
@@ -99,86 +98,86 @@ func toComDetails(debugPod *debug.DebugPod, ssOutput []string, protocol string, 
 	return res, nil
 }
 
-func identifyContainerForPort(debugPod *debug.DebugPod, ssEntry string) (string, error) {
-	pid, err := extractPID(ssEntry)
-	if err != nil {
-		return "", err
-	}
+// func identifyContainerForPort(debugPod *debug.DebugPod, ssEntry string) (string, error) {
+// 	pid, err := extractPID(ssEntry)
+// 	if err != nil {
+// 		return "", err
+// 	}
 
-	containerID, err := extractContainerID(debugPod, pid)
-	if err != nil {
-		return "", err
-	}
+// 	containerID, err := extractContainerID(debugPod, pid)
+// 	if err != nil {
+// 		return "", err
+// 	}
 
-	res, err := extractContainerInfo(debugPod, containerID)
-	if err != nil {
-		return "", err
-	}
+// 	res, err := extractContainerInfo(debugPod, containerID)
+// 	if err != nil {
+// 		return "", err
+// 	}
 
-	return res, nil
-}
+// 	return res, nil
+// }
 
-func extractContainerInfo(debugPod *debug.DebugPod, containerID string) (string, error) {
-	type ContainerInfo struct {
-		Containers []struct {
-			Labels struct {
-				ContainerName string `json:"io.kubernetes.container.name"`
-				PodName       string `json:"io.kubernetes.pod.name"`
-				PodNamespace  string `json:"io.kubernetes.pod.namespace"`
-			} `json:"labels"`
-		} `json:"containers"`
-	}
-	containerInfo := &ContainerInfo{}
-	cmd := fmt.Sprintf("crictl ps -o json --id %s", containerID)
+// func extractContainerInfo(debugPod *debug.DebugPod, containerID string) (string, error) {
+// 	type ContainerInfo struct {
+// 		Containers []struct {
+// 			Labels struct {
+// 				ContainerName string `json:"io.kubernetes.container.name"`
+// 				PodName       string `json:"io.kubernetes.pod.name"`
+// 				PodNamespace  string `json:"io.kubernetes.pod.namespace"`
+// 			} `json:"labels"`
+// 		} `json:"containers"`
+// 	}
+// 	containerInfo := &ContainerInfo{}
+// 	cmd := fmt.Sprintf("crictl ps -o json --id %s", containerID)
 
-	out, err := debugPod.ExecWithRetry(cmd, interval, duration)
-	if err != nil {
-		return "", err
-	}
+// 	out, err := debugPod.ExecWithRetry(cmd, interval, duration)
+// 	if err != nil {
+// 		return "", err
+// 	}
 
-	err = json.Unmarshal(out, &containerInfo)
-	if err != nil {
-		return "", err
-	}
-	if len(containerInfo.Containers) != 1 {
-		return "", fmt.Errorf("failed extracting pod info, got %d results expected 1. got output:\n%s", len(containerInfo.Containers), string(out))
-	}
+// 	err = json.Unmarshal(out, &containerInfo)
+// 	if err != nil {
+// 		return "", err
+// 	}
+// 	if len(containerInfo.Containers) != 1 {
+// 		return "", fmt.Errorf("failed extracting pod info, got %d results expected 1. got output:\n%s", len(containerInfo.Containers), string(out))
+// 	}
 
-	containerName := containerInfo.Containers[0].Labels.ContainerName
+// 	containerName := containerInfo.Containers[0].Labels.ContainerName
 
-	return containerName, nil
-}
+// 	return containerName, nil
+// }
 
-func extractContainerID(debugPod *debug.DebugPod, pid string) (string, error) {
-	cmd := fmt.Sprintf("cat /proc/%s/cgroup", pid)
-	out, err := debugPod.ExecWithRetry(cmd, interval, duration)
-	if err != nil {
-		return "", err
-	}
-	re := regexp.MustCompile(`crio-([0-9a-fA-F]+)\.scope`)
+// func extractContainerID(debugPod *debug.DebugPod, pid string) (string, error) {
+// 	cmd := fmt.Sprintf("cat /proc/%s/cgroup", pid)
+// 	out, err := debugPod.ExecWithRetry(cmd, interval, duration)
+// 	if err != nil {
+// 		return "", err
+// 	}
+// 	re := regexp.MustCompile(`crio-([0-9a-fA-F]+)\.scope`)
 
-	match := re.FindStringSubmatch(string(out))
+// 	match := re.FindStringSubmatch(string(out))
 
-	if len(match) < 2 {
-		return "", fmt.Errorf("container ID not found node:%s  pid: %s", debugPod.NodeName, pid)
-	}
+// 	if len(match) < 2 {
+// 		return "", fmt.Errorf("container ID not found node:%s  pid: %s", debugPod.NodeName, pid)
+// 	}
 
-	containerID := match[1]
-	return containerID, nil
-}
+// 	containerID := match[1]
+// 	return containerID, nil
+// }
 
-func extractPID(input string) (string, error) {
-	re := regexp.MustCompile(`pid=(\d+)`)
+// func extractPID(input string) (string, error) {
+// 	re := regexp.MustCompile(`pid=(\d+)`)
 
-	match := re.FindStringSubmatch(input)
+// 	match := re.FindStringSubmatch(input)
 
-	if len(match) < 2 {
-		return "", fmt.Errorf("PID not found in the input string")
-	}
+// 	if len(match) < 2 {
+// 		return "", fmt.Errorf("PID not found in the input string")
+// 	}
 
-	pid := match[1]
-	return pid, nil
-}
+// 	pid := match[1]
+// 	return pid, nil
+// }
 
 func filterStrings(filterOutFn func(string) bool, strs []string) []string {
 	res := make([]string, 0)
@@ -193,7 +192,7 @@ func filterStrings(filterOutFn func(string) bool, strs []string) []string {
 	return res
 }
 
-func parseComDetail(debugPod *debug.DebugPod, ssEntry string) (*types.ComDetails, error) {
+func parseComDetail(ssEntry string) (*types.ComDetails, error) {
 	serviceName, err := extractServiceName(ssEntry)
 	if err != nil {
 		return nil, err
