+++ pkg/types/types.go
@@ -13,8 +13,8 @@ import (
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/utils/ptr"
 
-	"github.com/liornoy/node-comm-lib/internal/consts"
-	"github.com/liornoy/node-comm-lib/internal/nftables"
+	"github.com/liornoy/node-comm-lib/pkg/consts"
+	"github.com/liornoy/node-comm-lib/pkg/nftables"
 )
 
 type ComMatrix struct {
